Include a fix for alternate login forms.
In situations where login_header() is inaccessible (because we’re on a
page other than wp-login.php because of a plugin or something) manually
redefine `login_header()` so we can call it there as well.

Also make sure we’re always submitting to the correct page and links go
to the correct page — `wp_login_url()` — if a plugin changes that, they
should also be reordering things to call the actions we use so that it
still works.

Resolves issues with WooCommerce logins from front-end pages. :+1:

Fixes #61