Make 'Log In' button default on verification form
On the Verification Code screen, currently if you
press return after typing/pasting the code,
'Resend Code' is the submit button that is triggered,
rather than 'Log In', which is confusing for users
as nothing appears to happen when they've provided
the correct code, yet they see ever growing incoming
emails (in the case of email verification).  For
most users Log In would be the expected behaviour.

Similarly, when you tab, Resend Code is the first
button rather than login.

Fixed by changing the order of the buttons in the
HTML (this also makes buttons horizontally aligned).