Fix Jetpack SSO Compatibility
After SSO, we successfully see the two factor verification. The problem is that the query string for SSO is like `?action=jetpack-sso&result=success&user_id=<redacted>&sso_nonce=<redacted>`. If we try to redirect back to that page, we get into a loop where SSO tries to login again, even though we're already logged in (and the nonce is no longer valid). We can redirect to /wp-admin/ instead of `REQEUST_URI`.