Include BaconQrCode library at version 1.0.3
The latest versions of this library do not support PHP 5.6 so we're
using version 1.0.3, the last release of the 1.x branch.

Command used to clone this code:

```
git clone https://github.com/Bacon/BaconQrCode.git
git -C BaconQrCode switch --detach 1.0.3
cp -r BaconQrCode/src/BaconQrCode includes
```

Fortunately the .php files in the library contain documentation on the
files' origin and licence so we do not need to add that information.

Link: https://github.com/Bacon/BaconQrCode/tree/1.0.3